ref(more): create standalone more button
