+++ spot-client/src/spot-remote/ui/components/index.js
@@ -1,5 +1,6 @@
 export * from './dial-pad';
 export * from './electron-desktop-picker';
 export * from './meeting-name-entry';
+export * from './more';
 export * from './nav';
 export * from './remote-control-menu';
