+++ spot-client/src/spot-remote/ui/components/more/index.js
@@ -1,0 +1,1 @@
+export { default as MoreButton } from './MoreButton';
