+++ spot-client/src/spot-remote/ui/components/nav/buttons/index.js
@@ -1,3 +1,2 @@
 export { default as AudioMuteButton } from './audio-mute-button';
-export { default as TileViewButton } from './TileViewButton';
 export { default as VideoMuteButton } from './video-mute-button';
